+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/dichotomy/DichotomyRunner.java
@@ -108,8 +108,8 @@ public class DichotomyRunner {
             return noCtStepResult.getRaoResult();
         } else {
             // initial network not secure, try worst case maximum counter trading
-            double ctFrEsMax = getMaxCounterTrading(ctRaFrEs, ctRaEsFr, expFrEs0, expEsFr0, "FR-ES");
-            double ctPtEsMax = getMaxCounterTrading(ctRaPtEs, ctRaEsPt, expPtEs0, expEsPt0, "PT-ES");
+            double ctFrEsMax = getMaxCounterTrading(ctRaFrEs, ctRaEsFr, expFrEs0, "FR-ES");
+            double ctPtEsMax = getMaxCounterTrading(ctRaPtEs, ctRaEsPt, expPtEs0, "PT-ES");
 
             double ctPtEsUpperBound = noCtStepResult.isPtEsCnecsSecure() ? 0 : ctPtEsMax;
             double ctFrEsUpperBound = noCtStepResult.isFrEsCnecsSecure() ? 0 : ctFrEsMax;
@@ -175,12 +175,12 @@ public class DichotomyRunner {
         }
     }
 
-    double getMaxCounterTrading(CounterTradeRangeAction ctra1, CounterTradeRangeAction ctra2, double initialExchange1, double initialExchange2, String borderName) {
-        double ctMax = initialExchange1 >= 0 ? Math.min(Math.min(-ctra1.getMinAdmissibleSetpoint(initialExchange1), ctra2.getMaxAdmissibleSetpoint(initialExchange2)), initialExchange1)
-            : Math.min(Math.min(ctra1.getMaxAdmissibleSetpoint(initialExchange1), -ctra2.getMinAdmissibleSetpoint(initialExchange2)), -initialExchange1);
+    double getMaxCounterTrading(CounterTradeRangeAction ctraTowardsES, CounterTradeRangeAction ctraFromES, double initialExchangeTowardsES, String borderName) {
+        double ctMax = initialExchangeTowardsES >= 0 ? Math.min(Math.min(-ctraTowardsES.getMinAdmissibleSetpoint(initialExchangeTowardsES), ctraFromES.getMaxAdmissibleSetpoint(-initialExchangeTowardsES)), initialExchangeTowardsES)
+            : Math.min(Math.min(ctraTowardsES.getMaxAdmissibleSetpoint(initialExchangeTowardsES), -ctraFromES.getMinAdmissibleSetpoint(-initialExchangeTowardsES)), -initialExchangeTowardsES);
 
-        if (ctMax != Math.abs(initialExchange1)) {
-            LOGGER.warn("Maximum counter trading " + borderName + " '{}' is different from initial exchange " + borderName + " '{}' ", ctMax, Math.abs(initialExchange1));
+        if (ctMax != Math.abs(initialExchangeTowardsES)) {
+            LOGGER.warn("Maximum counter trading " + borderName + " '{}' is different from initial exchange " + borderName + " '{}' ", ctMax, Math.abs(initialExchangeTowardsES));
         }
 
         return ctMax;
